Algorithms

Quac.Algorithms.QFT — Function

QFT(n)

Generate a Quantum Fourier Transform circuit of n qubits.

Quac.Algorithms.QuantumVolume — Function

QuantumVolume(n, depth)

Generate a Quantum Volume circuit of n qubits and depth layers. See [1] for more details.

References

[1] Cross, Andrew W., et al. "Validating quantum computers using randomized model circuits." Physical Review A 100.3 (2019): 032328.