Gates

In Quac, gates are symbolic, i.e. they do not store their representation. A gate instance just stores the qubit lane in which it acts and its parameters if needed. Thanks to Julia's multiple-dispatch different representations can be queried lazily from type information.

For example, this is a $Z$ that acts on qubit 4.

julia> gate = Z(4)
Z(4)

Any gate can be represented by a dense matrix.

julia> Matrix(gate)
2×2 Matrix{ComplexF32}:
 1.0+0.0im   0.0+0.0im
 0.0+0.0im  -1.0+0.0im

You can even specify the eltype!

julia> Matrix{Int}(gate)
2×2 Matrix{Int64}:
 1   0
 0  -1

Furthermore, the $Z$ gate allows a Diagonal representation!

julia> Diagonal{Float32}(gate)
2×2 Diagonal{Float32, Vector{Float32}}:
 1.0    ⋅
  ⋅   -1.0

Quac.Operator — Type

Operator

Parent type of quantum operators.

Quac.Gate — Type

Gate{Operator}(lanes...; parameters...)

An Operator located at some lanes and configured with some parameters.

Pauli gates

Quac.I — Type

I(lane)

The $σ_0$ Pauli matrix gate.

Note

Due to name clashes with LinearAlgebra.I, Quac.I is not exported by default.

Quac.X — Type

X(lane)

The $σ_1$ Pauli matrix gate.

Quac.Y — Type

Y(lane)

The $σ_2$ Pauli matrix gate.

Quac.Z — Type

Z(lane)

The $σ_3$ Pauli matrix gate.

Hadamard gate

Quac.H — Type

H(lane)

The Hadamard gate.

Phase gates

Quac.S — Type

S(lane)

The $S$ gate or $\frac{π}{2}$ rotation around Z-axis.

Quac.Sd — Type

Sd(lane)

The $S^\dagger$ gate or $-\frac{π}{2}$ rotation around Z-axis.

Quac.T — Type

T(lane)

The $T$ gate or $\frac{π}{4}$ rotation around Z-axis.

Quac.Td — Type

Td(lane)

The $T^\dagger$ gate or $-\frac{π}{4}$ rotation around Z-axis.

Rotation gates

Quac.Rx — Type

Rx(lane, θ)

The $\theta$ rotation around the X-axis gate.

Quac.Ry — Type

Ry(lane, θ)

The $\theta$ rotation around the Y-axis gate.

Quac.Rz — Type

Rz(lane, θ)

The $\theta$ rotation around the Z-axis gate.

Notes

The U1 gate is an alias of Rz.

Quac.Rxx — Type

Rxx(lane1, lane2, θ)

The $\theta$ rotation around the XX-axis gate.

Quac.Ryy — Type

Ryy(lane1, lane2, θ)

The $\theta$ rotation around the YY-axis gate.

Quac.Rzz — Type

Rzz(lane1, lane2, θ)

The $\theta$ rotation around the ZZ-axis gate.

General U2, U3 gates

Quac.U2 — Type

U2(lane, ϕ, λ)

The $U2$ gate.

Quac.U3 — Type

U3(lane, θ, ϕ, λ)

The $U3$ gate.

Controlled gates

Quac.Control — Type

Control(lane, op::Gate)

A controlled gate.

SWAP gate

Quac.Swap — Type

Swap(lane1, lane2)

The SWAP gate.